Add confirm delete modals

+++ app/instrument/challenge/video/controller.js
@@ -3,7 +3,12 @@ import Ember from 'ember';
 export default Ember.Controller.extend({
   actions: {
     deletePost(post) {
-      post.destroyRecord();
+      if(confirm('Are you sure you want to delete this post?')) {
+        post.destroyRecord();
+      }
+      else {
+        this.transitionToRoute('instrument.challenge.video')
+      }
     }
   }
 });
